requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
dataclasses==0.8
Django==3.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MetaTrader5==5.0.37
mysql-connector-python==8.0.31
mysqlclient==2.0.1
numpy==1.19.5
oauthlib==3.2.1
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
silence-tensorflow==1.2.1
six==1.15.0
sqlparse==0.4.2
support-developer==1.0.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
